he-weather/requirements.txt
click
helium-python
requests
apscheduler
